// src/components/ArcGIS/types/serviceArea.ts
export type ServiceArea = {
    OBJECTID: string
    code_service_area_id: number
    service_area_name: string
    watershed_manager: string
    cities: string
    construction_managers: string
    construction_counties: string
    area_total_acres: number
    area_total_sqmi: number
    CountAttachments: number
    AttachID: number
    created_user: string
    created_date: string
    last_edited_user: string
    last_edited_date: string
    Shape__Area: number
    Shape__Length: number
}